+++ Menu/Pages/ArenaMainLobbyPage.cs
@@ -156,10 +156,7 @@ public class ArenaMainLobbyPage : PositionedMenuObject, IDynamicBindHandler
             matchSettingsTab,
             new Vector2(120f, 0f),
             Arena.currentGameMode,
-            [.. Arena.registeredGameModes.Keys.Select(v => new ListItem(v, menu.Translate(v)))],
-            menu?.manager?.rainWorld?.inGameTranslator.currentLanguage == InGameTranslator.LanguageID.French
-                ? 270f
-                : default
+            [.. Arena.registeredGameModes.Keys.Select(v => new ListItem(v, menu.Translate(v)))]
         );
         arenaSettingsInterface.CallForSync();
         matchSettingsTab.AddObjects(arenaSettingsInterface);